Move autofill implementation to designsystem library
